refactor: extract useLeaseOpen from Long/Short forms (#185)
LongForm and ShortForm carried byte-identical refs, input handlers, and
the three downpayment validators (the reactive insufficient-balance check
is the #192 fix). Centralizing that infrastructure in useLeaseOpen removes
the duplication that let the two forms drift.

The divergent, financially load-bearing parts stay in each form on purpose:
assets/coinList building and the leaseTicker / leaser-protocol derivation in
calculate()/openLease() — Long reads the ticker from the loan key, Short
resolves the protocol stable (group="lease") async. Keeping those per-form
preserves the Long-vs-Short asymmetry where a reviewer can see it, rather
than hiding it behind a shared strategy. The validators take the resolved
downpayment/loan options as arguments so the composable never needs the
per-form asset lists.
